Co-locate map update check and its backing logic
